DRY out build/destroy subcommands.
